Fix the ordering of the libc++.so linker script.
libc++abi and libandroid_support are static libraries, so we need to
be careful about ordering them. They depend on libunwind and
libatomic, so we need to make sure those come after.

For some reason the tests in r12 tripped this and this branch didn't.
Might be something related to the libc++ update that we have in
master?

Bug: http://b/29275056
Change-Id: I60bbf6239fb75430cdf3bf9599e3a6bccc5332b7
(cherry picked from commit cf0ae5010c1597ee17ec0637579bd0b67754c63e)