Add APP_STRIP_MODE and LOCAL_STRIP_MODE.
These arguments allow the user to specify how their binaries should
be stripped. APP_STRIP_MODE indicated what to do if LOCAL_STRIP_MODE
is not set. LOCAL_STRIP_MODE always overrides APP_STRIP_MODE.

The value of this argument is passed directly to the strip command, so
valid values include anything accepted by strip, but commonly will be
--strip-debug or --strip-unneeded.

A special value may be used to indicated that the binary should not
be stripped at all: none. This is helpful for users that want to
preserve even debug info, perhaps during debugging or to delegate the
stripping behavior to their packaging system (Studio does this).

Test: ./run_tests.py --filter strip*
Bug: https://github.com/android-ndk/ndk/issues/749
Change-Id: Ia0748cb7b9b92bcbc164f5156a6ad0e523314be2