Fix packaging of the sysroot on Windows/Mac.
linux/netfilter has some headers with names that differ only by case,
which can't be extracted to a case-insensitive filesystem, which are
the defaults for Darwin and Windows :(

There isn't really a good way to decide which of these to keep and
which to remove. The capitalized versions expose different APIs, but
we can't keep both. So far no one has filed bugs about needing either
API, so let's just dedup them consistently and we can change that if
we hear otherwise.

Test: ./checkbuild --module sysroot --system windows64 && \
          unzip -l ../out/dist/sysroot.zip | sort -f | uniq -di
Bug: None.

Change-Id: Ib4bba4fc4e6d1e1fe803040649c2f61bb6d5cdb1