Warn standalone toolchain uses default API level.
Some people expect this to default to the latest, some expect the
earliest, others expect it to be a required argument. The latter
seems like the best choice, but we're stuck because it's historical
and people don't like it when we change things, so just print a
warning.

Test: build/tools/make_standalone_toolchain.py --arch arm
Bug: https://github.com/android-ndk/ndk/issues/270
Change-Id: Idb4cd6d90eeabb788778540e52c2c2eb809308cc
(cherry picked from commit e6e7857c57ce35f9fae69510926b954d2dafccad)