Partial revert of mips64 toolchain for 32-bit.
https://android-review.googlesource.com/c/247064/

We need to drop this because neither Clang nor GCC is able to
correctly find the gnustl headers if only the 32-bit ones are present.

This might be worth revisiting at some point, but for now I'm just
moving standalone toolchains back to the 32-bit toolchain to get
things back into working condition. This change is working fine for
cmake and ndk-build though, so it's only a partial revert.

Tests are in the next commit so we don't accidentally revert the
tests if we revisit this.

Test: ./validate.py --filter standalone_toolchain_gnustl*
Bug: https://github.com/android-ndk/ndk/issues/310
Change-Id: I2df811c5a67db9a64d3d352e08d60f660e075012
(cherry picked from commit 40d8abbf40e0556ccfa5598ff8c4ac8c1e141934)