ndk-stack: link in the libz packaged with binutils.
This fixes the ndk build for windows that has been failing
with link errors due to "undefined reference to `inflate'".

binutils-2.27 adds support for compressed debug sections,
and as a consequence of that, libbfd.a now has a dependency
on libz.a. The following upstream commit made linking libz
mandatory in order to use libbfd:
  - https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=243340ad57e30de3c15cffe6e4af0381995f15d9

Bug: 65591283
Test: './ndk/checkbuild.py' passed.
Test: './ndk/checkbuild.py --system=windows' successfully builds ndk-stack.

Change-Id: Ia5b0f767ee739cabf59e56bfda7ddf4cf6bebcd0
(cherry picked from commit f659751bf167420930bd558ca958c558b625cac8)