Remove the local hack to the clang prebuilt.
clang-4053586 was patched in the prebuilts directory to add the libc++
includes. These are almost certainly a different revision than the NDK
libc++, and may contain local changes that the NDK's don't and vice
versa. Best to just remove them for the time being since that returns
to the previous behavior.

The test doesn't check ndk-build. For CMake, this should work because
cross-platform code may be using this. ndk-build is specific to the
NDK, and this flag shouldn't be used. It doesn't currently work as an
ldflag because the compiler will attempt to use `-lc++` and won't find
a library to satisfy those requirements since we don't put the STL
directory on the library path.

Test: ./run_tests.py --filter stdlib-libcxx
Bug: https://github.com/android-ndk/ndk/issues/564
Change-Id: I37076b838159a97cc2064cade49a6990287d5b34
(cherry picked from commit 5c403c69674e22f1395dc61998bb7217016bd751)