Don't use libandroid_support for modern API levels.
If we're targeting a new enough platform version, we don't actually
need to cover any gaps in libc for libc++ support. In those cases,
save size in the APK by avoiding libandroid_support.

This is also a requirement for static executables, since using
libandroid_support with a modern libc.a will result in multiple symbol
definition errors.

Test: ./checkbuild.py && ./run_tests.py
Bug: https://github.com/android-ndk/ndk/issues/272
Change-Id: Iec548d7a75cc407d091b1a2bc632df68756c717d