Fix link ordering for exceptions with system STL.
LOCAL_LDLIBS happens last, and we need to make sure that libunwind
comes before libgcc. This needs to happen before
modules-compute-dependencies is called so the dependency graph will
still be correct. Move it into modules-add-c++-dependencies.

The test that would have caught this were accidentally removed as part
of the larger remove gnustl/stlport cleanup. Revive it and the related
test for RTTI (renamed to be nice to Python).

Test: ./run_tests.py --rebuild --filter system_stl_*
Bug:
Change-Id: If74272f33fda7d65b9535926a0377c04615e1ab4