Fix case of APP_WRAP_SH being empty.
I used the wrong variable to check the sanitizer flags, so it was
always using ASAN, causing the asan wrap.sh to be installed unless
the user specified their own.

Fix the problem and add a test case to make sure this is only
installed if it's requested.

Also clean up some debug logging that was accidentally left in.

Test: ./run_tests.py --rebuild
Bug: https://github.com/android-ndk/ndk/issues/540
Change-Id: I0e6d4b0222a4aeb1b491f6869c338011b74e2e8f