Cleanup handling of LLVM_TRIPLE.
Right now arm32 is the only ABI that appends the API level to the
triple, and it doesn't do it correctly. For one, it does so twice
causing the triple to be something along the lines of
arm-linux-androideabi1616 at exit from the file, though it's not
currently used outside this file. Second, it should be using
TARGET_PLATFORM_LEVEL, not APP_PLATFORM_LEVEL. The latter does not
get adjusted if it's below the minimum.

We should move all ABIs over to using this, but for now make it
consistent with the other ABIs by skipping it.

Test: ./checkbuild.py
Bug: None
Change-Id: Id69335c706645cfd46cd170f3429c714b0fe99ab