Fix LP64 standalone toolchains targeting 21+.
We were using the same libc++ linker script regardless of API level,
and for 21+ we don't copy libandroid_support since it is not needed.

In this case, rewrite the linker script after we copy it into the
standalone toolchain. We should investigate a different fix for r18,
but I don't think we should be making a more disruptive change this
late in the r17 release.

Test: ./run_tests.py --rebuild
Bug: https://github.com/android-ndk/ndk/issues/672
Change-Id: Ia3db77194953c5b72b12b6b4a9320444e2df3ab3