Add meta/platforms.json.
Gradle (as well as a few of our own tools, but those can more easily
change in sync with the NDK) needs to know what API levels a given
NDK supports. It currently does this by scraping the $NDK/platforms
directory, but that directory will be removed in r20 (or perhaps later
depending on how long migration to the new toolchain layout takes).
Include this information in meta/platforms.json as a sanctioned source
for this information.

Test: ./checkbuild.py meta
Bug: None
Change-Id: Id5c59dff048e3dd510c3f315a1d18971541078b6