Ship lld.
This is looking good enough to start shipping as experimental. The
only thing that needs investigating right now is whether we need to
add support for --fix-cortex-a8. LLD doesn't support the errata fix
right now, but we need to figure out if any Cortex-A8 devices are
still out there before before we add support for that.

Test: ./run_tests.py --rebuild # after sprinkling -fuse-ld=lld into
                               # the various build systems
Bug: https://github.com/android-ndk/ndk/issues/683

Change-Id: If14db8bb1d037f99293d41f85040f719b87731d2