Move tests.testlib into modules in ndk/.
Finally have all of our Python code in the proper place... all the
Python files still in the tests/ directory are test_config.py or
test.py files that are part of the actual tests.

The only non-trivial code change here is that the old
tests.testlib.TestRunner has been merged into
ndk.test.builder.TestBuilder. The latter was essentially a more
convenient API for the former anyway.

Test: ./checkbuild.py && ./run_tests.py
Bug: None
Change-Id: I620c790b6d5f7984d89704267e9e771528e1ff66