Build GCC prebuilts with x18 reserved.
Once https://reviews.llvm.org/D45588 lands in our copy of Clang we
should get x18 reserved in everything built by Clang, but that won't
cover the prebuilts built by GCC (most notably libgcc.a). This CL
changes the cflags and cxxflags that we pass to GCC so that libgcc.a
is built with x18 reserved.

Bug: 112907825
Test: cd toolchain/gcc && python build.py --toolchain aarch64-linux-android
Change-Id: I9d40356dae89ca24fbea34c89acdf86a1cc28b2d