Build the Linux NDK alongside Windows.
We want to migrate to building libc++ (and similar components) using
the just-built NDK. To do this, we need to have a just-built NDK that
matches the current host rather than the host we're building for.
Build the Linux NDK before building the Windows NDK to
facilitate this.

To handle this, there is now an out directory per host (i.e.
out/linux, out/windows64). A symlink will be created linking
out/android-ndk-rN-canary to the NDK for the given host since that's
how we're accustomed to accessing it, but each NDK will be installed
in its host directory.

When Windows is targeted, Linux dist artifacts (intermediate
artifacts that really shouldn't be in dist any more, plus logs) are
installed to the Linux out directory rather than out/dist, since that
need only contain Windows artifacts.

As it is much more common to build a Linux NDK than a Windows NDK for
local development, the Windows build will reuse the existing Linux
NDK if available. `--skip-deps` will cause the Linux NDK to not be
built (note that deps need to be built the first time the Windows NDK
is built so the Windows NDK can build its own deps).

Test: ./checkbuild.py --system linux
Test: ./checkbuild.py --system windows64
Test: ./checkbuild.py --system windows64 --skip-deps ndk-build
Bug: None
Change-Id: I81c87bde928d80bb01c9e0aaff9a8937e3177ad1