Refactor the build/install module interface.
This change factors out out_dir, dist_dir, and args into a
BuildContext object. This allows us to clean up many of the
path-related interfaces in the build system.

This also makes it easier for us to guarantee safe access to
dependencies. Since modules now need some initialization, we can no
longer trivial construct a new instance of a module to make use of
it. Instead, we go through the module list contained by the
BuildContext object and gate that access to only modules listed in
the module's dependency list.

Test: ./checkbuild.py
Bug: None
Change-Id: Idbf650e4606c0e370aaa6ac3393296b55d95e6a5