Improve generate-list-file performance.
Using list file generation for creating compile_commands.json slowed
down ndk-build a fair amount for large projects. For the following
command:

    $NDK/ndk-build -C $NDK/sources/third_party/shaderc \
        NDK_PROJECT_PATH=. APP_BUILD_SCRIPT=Android.mk APP_ABI=all \
        APP_PLATFORM=android-24 -j 72 clean

r17b takes 0.5 seconds, but r18 takes 10 seconds.

The algorithm itself is rather slow and can be sped up. With this
change, the above command takes only 0.9 seconds.

We should still gate the use of this on LOCAL_SHORT_COMMANDS so we
can get back to r17's speeds, but that will be a follow up patch.

Test: Generated a compile_commands.json for a project and compared it
      to one generated with r18. The only differences are whitespace.
Bug: https://github.com/android-ndk/ndk/issues/799

Change-Id: Ic8f58c8f2c05c19b613e095c0479049b31ad6d72