Only use list files for the JSON db when requested.
Generating list files is faster now, but just the generation of the
rules is still quite slow when there are many objects. Avoid doing so
unless LOCAL_SHORT_COMMANDS is used.

When the following command is used to build a project with 1000 empty
source files:

    ndk-build APP_ABI=arm64-v8a -B -j72 clean

r17b takes 0.25 seconds and r18 takes 11 seconds.

With this change, r19 takes 1 second when LOCAL_SHORT_COMMANDS is
true and .32 seconds when it is false.

Test: Generated a compile_commands.json file and checked accuracy
      with both LOCAL_SHORT_COMMANDS=true and
      LOCAL_SHORT_COMMANDS=false.
Bug: ndk-build -C foo APP_ABI=arm64-v8a -B -j72 clean

Change-Id: I5a440505c68d52f30131290998a5ce080f489732