Bootstrap checkbuild.py with Python 3.
This allows us to implement checkbuild.py and its associated pieces
exclusively in Python 3 even though it isn't available on the build
servers. The version currently available in external/python/cpython3
is 3.6, so this gets us features even as new as type annotations and
f-strings.

On my P920 the bootstrapping step takes about a minute. When
bootstrapping is complete a .bootstrapped file is placed in the out
directory to avoid rebuilding Python every time.

We can't go through our existing Python 2 as an intermediate step
because we can't build the SSL (and therefore the hashlib) module for
that version of Python because it's not compatible with whatever is
on these machines (which might be BoringSSL?).

Test: ./checkbuild.py
Bug: None
Change-Id: Ie01a586cdbfaf22da3025aabbc3ed931480b7742