Switch to llvm-strip, advocate its use.
Using binutils strip with lld breaks RelRO. Switch to llvm-strip to
avoid this issue.

This doesn't affect Gradle users because the Android Gradle Plugin
performs its own stripping. Gradle users will need to upgrade to
AGP 3.6 or newer (the change to fix the plugin has not yet been
submitted, but is in review).

Test: ./checkbuild.py && ./run_tests.py
Test: Compare sizes of all test outputs before and after, no changes
Bug: https://github.com/android-ndk/ndk/issues/843
Change-Id: I6710d0bc5dbbdc53bc3a66acccc6ba3d9c9a00f6