Refactor test config inputs.
Alter the inputs to test config functions so they take the test and
(for the run phase) the device rather than the whole test matrix.
This makes it so we don't have to modify every test config function
whenever we need to add a dimension to the test configuration (such
as lld).

Test: ./checkbuild.py && ./run_tests.py
Bug: None
Change-Id: If2b9f34b0898bcc06251894535e0a7eefd74be6b