Reorganize the roadmap.
Now that much of the upcoming work is not inherently tied to an NDK
release (cdep improvements, as an example), it's not as useful to be
slotting our work directly into a given NDK release. Move the major
work down to its own section in a rough priority order to decouple
them and leave the release roadmaps to cover only the things we know
will be a part of a specific release.

Test: None
Bug: None
Change-Id: I98068ee0732372779c4805f7013e56d64950c49f