Disable multithreaded linking for Windows LLD.
This only affects ndk-build because CMake can't currently check if it
is using LLD since our toolchain file has finished running by the
time the user can add ldflags.

Not enabling this with standalone toolchains since we can't enable it
for the built in toolchains without recompiling Clang and those are
supposed to match. Will disable in the driver in a follow up if a
workaround is not found.

Test: ./checkbuild.py
Bug: https://github.com/android-ndk/ndk/issues/855
Change-Id: I4a24166adce463bea47ca4bf0e97d5375c69f40e
(cherry picked from commit 144b2449de418b486b53d35a0d9916b75c7d5265)