Fix ARM version for non-integrated assembler.
The non-integrated assembler does not get the proper -march flag when
it is implicit in the triple. Pass it explicitly to avoid problems
until we can fix the driver.

Writing a test that fails to compile is proving difficult because the
warnings emitted by the assembler for the provided test case are just
notes rather than warnings so they can't be promoted to errors, and we
can't use the __ARM_ARCH_7A__ define because those are resolved by the
C preprocessor, which *does* have the right architecture set...

Will attempt to add tests for this in a follow up since we need to
respin r19b ASAP.

Test: Manual inspection for now.
Bug: https://github.com/android-ndk/ndk/issues/906
Change-Id: I4ab3e9aee0775150731108e42f5f584a5f3c27ea
(cherry picked from commit e3f70782abf6e4567aa1cfc025b001d3c125f06f)