Stop making repo.prop files.
We don't make these for every project and aren't actually necessary.
The build number is all that is needed to determine the pedigree of
the NDK.

Not removing the make_repo_prop API in this commit because I'm
guessing it's used by things like the GCC/binutils builds and I need
to clean up those callsites first.

Test: ./checkbuild.py && ./run_tests.py
Bug: None
Change-Id: I2347a4c31f8b2d40b0def94b90d45dc618ab5c08