Allow users to set their own CMAKE_FIND_ROOT_PATH.
If we set this option here we will clobber anything the user has
provided on the command line. Despite the name, this is actually a
list. All we need to ensure is that the variable is not empty, so
appending to the list is sufficient to avoid the host's libraries.

Test: ./checkbuild.py && ./run_tests.py
Bug: https://github.com/android-ndk/ndk/issues/912
Change-Id: I683a99c556e619e8fec86cbb89450407562e40b9