Fix timer doc test.
The syntax was wrong for the line continuation which was causing this
test to fail.

We don't really want to make this test flaky by having it depend on
the actual time spent running the test (though it should never take
more than 1 second, which is the minimum resolution of default
output), so hide it in a function so the block is still checked but
not run.

Test: nose2 ndk
Bug: None
Change-Id: Id41149cd27ab19a9c763db5ce45345742a559dd3