Fix bad Windows package name.
get_host_tag doesn't get the host tag for a host, it gets the host
tag associated with a given built NDK. The code paths that inspect
the argument only occur on a Windows host (from the test runner), and
this does return the host tag of the current host on Darwin and
Linux, so the behavior was correct there. My local incremental builds
also didn't detect this error because there was already a Windows
artifact in my out directory that could be packaged.

Test: rm -rf ../out && ./checkbuild.py --system windows
Bug: None
Change-Id: Ia59fafe9e725aaf961b847d23edde20b860e0cfa