Remove hard coded system_libs.json list, fix doc.
We didn't actually ship this. The proper list was being autogenerated
and this was only used for documentation purposes, but since it wasn't
tested it was usually not up to date. Remove it and just refer to the
location in the NDK with the real specification rather than a link to
a stale metadata file.

Test: ./checkbuild.py && cat $OUT/.../meta/system_libs.json
Bug: None
Change-Id: I8cb0472e22748333b408427b3378fc028bfb530b