Fix ndk-stack to work on Windows.
Looking for llvm-symbolizer on Windows doesn't work since it has
an .exe extension. Change the name of what's being looked for on
Windows.

This kind of worked before because it always fell through to the
second case and returns the default location. That works on the
normal setup since calling llvm-symbolizer without the exe extension
is okay.

Test: Ran this on linux machine with a stack trace and verified it
Test: produces valid file/line numbers.
Test: Ran this on mac machine with a stack trace and verified it
Test: produces valid file/line numbers.
Test: Ran this on windows machine with a stack trace and verified it
Test: produces valid file/line numbers.
Change-Id: Ie9986561bc12541fc5feda5265d1c1b3521e88ad