Allow externally specified import paths.
Gradle needs to be able to specify a dependency import path for Prefab
modules. Add NDK_GRADLE_INJECTED_IMPORT_PATH to support this.

This could instead be a more generic NDK_ADDITIONAL_IMPORT_PATHS, but
that's probably of limited use since Gradle will clobber the user's
flags with its own.

Test: ./run_tests.py --rebuild
Bug: https://github.com/android-ndk/ndk/issues/916
Change-Id: I18f9ff2830846d9ea63130d6949c9d199b72db60