Fix openmp tests.
We don't currently have any infrastructure for copying the openmp
runtime to the device as part of the test. Keep using the static
openmp like we used to.

For ndk-build, we should be copying the runtime library to the install
directory if appropriate.

For CMake, the install directory is not used. We have this same
problem with ASan, so if we decide that CMake should do anything about
that we can handle openmp the same way.

Test: ./run_tests.py --rebuild --filter openmp
Bug: https://github.com/android-ndk/ndk/issues/1028
Change-Id: Id41455c7dad1117aeae4587a2bc1b9601d55e472