Use $(file) from Make 4.
This looks to be about twice as fast and lets us delete some code. To
test performance, I generated an ndk-build project with 1000 empty C++
files and built a single static library from them. That project was
built with:

    $ time $NDK/ndk-build -C foo LOCAL_SHORT_COMMANDS=true -j 72 clean

Checked with r18b (the last time we improved list file performance),
and master with and without this patch.

The results:

NDK                  | Time
-------------------- | ----
r18b                 | 5.4s
r21 (without change) | 4.5s
r21 (with change)    | 2s

Test: ./checkbuild.py && ./run_tests.py
Bug: https://github.com/android/ndk/issues/799
Change-Id: If8b13cc72b462bda9e5238acd12f5e64f3bf0a96