Move __ANDROID_NDK__ to <android/ndk-version.h>.
We did try to move this to bionic's <android/api-level.h>, but that
doesn't work well because close to a release the platform build's API
level goes from 10000 to 29 or 30 or whatever, at which point you can't
distinguish "NDK" from "platform build but with `sdk_version` set".

The only way to truly know if you're building with the NDK is if the NDK
declares the fact.

(Also switch to `#pragma once`.)

Test: treehugger
Change-Id: I5d835f1f73de6080deab691d2a0d6a91b1b03291