Fix handling of crash dump data.
There was a bug where taking the data from a tombstone would result
in no actual output. Fix that.

Also, fix the .gitignore so that .so files can be submitted in to
the test directory.

Add missing shared library files and update input/output files to system
test.

Add a new system test that contains multiple full tombstone information.

Test: Unit tests pass.
Change-Id: I77ad59ba2c6a109cdbe11a5aae411d5b34f8daa5