Only use explicit --build-id=tree for lld.
Using sha1 instead of tree turned out to have a larger build impact
than the benchmarking suggested (10% regression in dolphin build
time).

For ndk-build, scan all the linker flags for -fuse-ld and check
APP_LD to determine if lld is being used, and if it is then fallback
to tree. Otherwise, use the linker default.

For CMake we don't know the user's chosen ldflags, so the best we can
do is work off of ANDROID_LD. If the user doesn't use ANDROID_LD and
instead uses -fuse-ld=lld explicitly, they won't be able to debug.

Test: inspected verbose build output for the following configurations:
    * No flags
    * APP_LD=lld
    * APP_LD=gold
    * APP_LD=lld APP_LDFLAGS=-fuse-ld=gold
    * APP_LD=gold APP_LDFLAGS=-fuse-ld=lld
    * For the first three variants, checked CMake with ANDROID_LD.
Bug: http://b/143411084
Bug: https://github.com/android/ndk/issues/885

Change-Id: I62a7b6a296a61bc543c38d406bfb9f15ddb1c7fa