Tone down the legacy toolchain removal note.
Removing *all* of these hinges on us being able to remove binutils
because a final part of the cleanup will be removing the
target-specific bin directories within the LLVM toolchain, which can
affect driver behavior. Since we don't know if we'll be able to get
that done for next year's LTS, don't be so explicit about the
timeline. We can be more specific when we know more.

Test: None, doc only
Bug: None
Change-Id: I0ddbd948c1c45b894d417b0ef3d1bfded0eccc7a
(cherry picked from commit f8286853cc7e6dd2cd802e571b8452853f3ae708)