Only create zip archives of the NDK.
We never gained the ability to distribute the tarballs like we wanted
to do for Linux/Darwin, and the SDK requires zips.

I'm changing the default behavior here as well. It takes almost 10
minutes to zip -9 the NDK even with these beefy machines since zip
isn't parallel. I don't want that to be part of our regular workflow,
so `--package` must now be explicitly given to cause packaging to
happen. The build servers already pass this argument (long ago we
didn't package by default, and I never got around to cleaning up the
build configs when I flipped the default the last time).

Test: ./checkbuild.py
Test: ./checkbuild.py --package
Bug: None
Change-Id: Iad8e6579f54135881d87fc0473ea746262da648e