Speed up the test runner.
At some point we started running the tests fast enough that at the end
of the test run we'd actually have finished running all the tests but
hadn't popped all the results yet. When we get into that situation, we
can actually spend a *long* time (10 minutes out of 17, on a run that
I just had) just redrawing the UI here for every result.

Pull all available results and redraw the UI *once*.

If we have time in the future, it'd probably be good to replace (or at
least reimplement) the work queues with concurrent.futures since we
have that available now. The API probably will change a bit to account
for that though, so this is enough for now.

Test: ./run_tests.py
Bug: None
Change-Id: If111b9a7457875753c960e080ceecd955a347a5c