Note std::filesystem support in the changelog.
The tests are not included yet because I need to send a bunch of
patches upstream and get a prebuilt toybox into the tree. I'll follow
up with those. I have all the patches locally and tests are passing
with them, so this is good enough to ship in r22.

Test: run_tests.py with a bunch of test patches
Bug: https://github.com/android/ndk/issues/609
Change-Id: I4e8a5e04907920bf42d643d393fa1ec5365f85a2