Install lld as bin/ld rather than bin/$TRIPLE-ld.
It seems Clang doesn't search for a triple prefixed ld executable, but
just a bare `ld` in either the driver directory or the triple-specific
bin directory. Fixing this makes the executable in the triple-specific
directory not used in our workflow  it seems, but we'll continue
maintaining that path until we remove the directory entirely when the
rest of GNU binutils is removed.

Test: ./checkbuild.py && ./run_tests.py
Bug: None
Change-Id: Ie466cd5b1e0000d288cc9ffd861db31fc20faec9