Clean up CMake compiler identification overrides.
The bug that prevented this from working for cross-compiles with Clang
was fixed at some point. Avoid overriding when we know we don't have
to so we don't accidentally introduce bugs with missing/wrong data.

To reduce maintenance for the old path, automatically generate those
values during the build.

We could potentially make this check a little better. I don't have any
builds between 3.6 and 3.10 available, so it's possible we don't need
to force these values in some versions between those, but this
preserves existing behavior for those versions, cleans up the only
part that we keep breaking, and the SDK doesn't include any
intermediate versions anyway. Anyone using a non-SDK version of CMake
is likely using something even newer than 3.10 rather than something
between those two.

With this fixed the only thing we need to do to update LLVM is pick
the new version in ndk.toolchains, making it possible for us to do
choose the version automatically for a chained build to test new LLVM
versions.

Test: ./checkbuild.py && ./run_tests.py
Test: Checked that these values were still being set appropriately
    with 3.6
Bug: None

Change-Id: I6c8d6b2c15e006aa7282343fb6e416260e0ee237