Enable branch-protection for crt objects
This change turns-on "-mbranch-protection=standard" for crt objects
targeting aarch64. This flag is required to emit .note.gnu.property
which signals if the ELF file supports
  - Armv8.3-A Pointer Authentication (PAuth) or
  - Armv8.5-A Branch Target Identification (BTI).

BTI compatibility is important because crt objects are linked to
virtually everything. Unless BTI compatibility is set in the
.note.gnu.property section no linked so-s and binaries will have
BTI compatibility enabled.

Source code of crt is already updated to support both of these features.
All emitted PAuth and BTI instructions are in the hint space and are
executed as nops on targets which don't support these extensions or
if the extensions are turned off.

Change-Id: I80cef8e190ca9b2c089413d004d91c2e09fff8bc