Remove non-functional test.
The `$API_LEVELS` variable that this test iterates over has been empty
for a while.

The test itself is valuable, but this is also too late to be running
it. This should be a part of the sysroot generation in the platform
build. To an extent we already have covereage for this: versioner
performs a better version of this check for bionic.

Test: treehugger
Bug: None
Change-Id: I9659ad7627e6ca288423fb6800b947d7de4674e8