Update release table generator for adrt changes.
We had to drop the host arch for Darwin because it prevent the SDK
manager from serving the NDK to aarch64 macs. We dropped it from all
artifacts for consistency.

The DMG artifact also had to move in adrt such that the NDK artifacts
are not all grouped in the lorry email, so skip non-NDK lines so the
whole table from the email (which may contain things other than the
NDK) can be pasted.

Test: Used this for r23 beta 5
Bug: http://b/189856862
Change-Id: Ie5e53f6bc62e24b2119c3eba16298727a5136eab