Fix test.py tests to support each CMake toolchain.
Change the interface to accept the build configuration instead of
individual components so we don't have to change the interface every
time BuildConfiguration changes.

This also starts running each CMake variant for these tests which were
previously filtered out during scanning. This is somewhat wasteful for
all the cases that don't use CMake, but that should be set in the
test_config.py instead. I haven't done that since these make up such a
small portion of testing time.

Test: run_tests.py --rebuild
Bug: None
Change-Id: I70fbcfaee4300fc2a93ba04612cd9002514e2646