Add update_kokoro_prebuilts.py script
The script accepts a list of Kokoro build IDs (a list of UUIDs). It
downloads each build's main archive and extracts it into the
appropriate place in an ndk-kokoro-main repo/pore checkout. It
automatically creates a branch and a commit updating each prebuilt.

The script uses the `stubby` CLI tool to access the Kokoro API to
retrieve build details. This gives it the location of Kokoro builds (in
placer) as well as a list of Git-on-Borg SHAs. For the given set of
build IDs, it verifies that no two builds use different Git SHAs for a
given repository, which guards against accidentally updating two hosts
to different versions.

The script uses `fileutil` to download artifacts from placer.

The script needs the google.protobuf Python module.

I generated kokoro_api_pb2.py by running:
protoc kokoro_api.proto --python_out .

Change-Id: Ibaa3ade7fb21ef69b29b985dd6164d3c8cdf878d