cmake: list ASM after C and CXX in project()
According to the CMake docs, ASM should come last so that CMake can
first look for the C/C++ compilers and see if they also handle ASM.

With ASM listed first, the test fails if CMake doesn't find an
assembler. e.g.:

    FAIL cmake_not_mingw [x86-19-new]:
    -- Android: Targeting API '19' with architecture 'x86', ABI 'x86', and processor 'i686'
    -- The ASM compiler identification is unknown
    -- Didn't find assembler
    -- Android: Selected unified Clang toolchain
    -- The C compiler identification is Clang 12.0.8
    -- The CXX compiler identification is Clang 12.0.8
    CMake Error at CMakeLists.txt:2 (project):
      No CMAKE_ASM_COMPILER could be found.

The test fails like this on Windows, as well as on an unmodified
centos:7 container (but passes if gcc is installed).

Bug: none
Test: ./run_tests.py --filter cmake_not_mingw --build-only
Change-Id: I47c35320c72f2067efe781d78a61e6292bdb0eac