Delete libc++.a and libc++abi.a on macOS.
With the switch to universal binaries in the LLVM prebuilt, these files
are no longer ar archives, but instead are Mach-O binaries that each
contain multiple ar archives. Our release staging infrastructure signs
these archives, but Apple's notarization service rejects these two
files in the archive we upload.

According to this thread[1], Apple's code signing tool stores the
signature in an (AFPS/HFS) extended attribute, and we need to use ditto
to preserve this attribute in a ZIP file, because "zip" discards it.
Switching to ditto might let us distribute these two files, but we
don't really need them, so it's easiest for us to remove them.

I've left them in for Windows and Linux, but maybe we should stop
distributing those too.

[1] https://developer.apple.com/forums/thread/126038

Bug: http://b/200995199
Test: checkbuild.py
Change-Id: Ic217d1c7bc3b24741f9c4d9c1e8209c5e173c5fa