Stop using CRLF for Windows shell script
Cygwin bash (and Ubuntu bash running under WSL) don't recognize the CR
in CRLF as part of the EOL, and instead treat the CR as the last
character in the line. (MSYS2 bash, on the other hand, does recognize
CRLF.)

A shell script should always use LF, even on Windows.

This change also removes the dos2unix dependency.

Bug: none
Test: checkbuild.py, manually run spirv-lesspipe.sh on a Windows Bourne
  shell

Change-Id: I607a3ae69805387c1e35d630d314ed74dc2aee30