Remove make-standalone-toolchain.sh
This script was broken by the Id321987c11e0f6a29c13fbca80e77967d56dafaf
change, which removed prebuilt-common.sh, which this script includes.
That change was shipped in r23, the current LTS.

This script is just a wrapper around make_standalone_toolchain.py,
which works with either Python 2 or Python 3.

Bug: none
Test: n/a
Change-Id: I7dc1f42a6e9b10881cf66fbc1463ec64ae43ba4e