Lint ndk-build and wrap.sh/asan.sh with shellcheck
It detected a bunch of places where a space wouldn't be handled.
ndk-build tries to report an error if it's run from a directory with a
space, but it only works if $0 has no space in it. (If it has a space,
then cd fails, so $PROGDIR is an empty string.)

Some of the variables in ndk-build are unused, e.g. as a result of
replacing scan-build with clang-tidy,
Ibca4fc7a9d9a37ef9ddc90446b2a6c0102a9125c.

Bug: none
Test: shellcheck build/ndk-build
Test: shellcheck wrap.sh/asan.sh
Test: checkbuild.py && run_tests.py
Change-Id: Ia7d404485f73d32a36c662889cdabed99d12e18d