Add missing CMAKE_ASM_COMPILER_VERSION fallback.
Since older versions of CMake can't get compiler ID detection for the
NDK right we need to specify these manually, and we forgot to get all
the languages. CMake uses the compiler version to decide whether the
spelling is -gcc-toolchain or --gcc-toolchain=. The latest LLVM
doesn't support -gcc-toolchain any more, so it's important that CMake
gets this right.

Test: updated LLVM, didn't see build failures for CMake asm tests
Bug: None
Change-Id: I20159e8e7a4fd2638c9f552d647e55720b16c76c