CMake legacy toolchain: stop passing -gcc-toolchain
The flag became unnecessary with the reorganization in r19 [1], which
also stopped passing the flag for ndk-build. The non-legacy CMake
toolchain file also omits the flag.

Newer CMake versions don't detect the Clang version for the ASM
language, so they incorrectly set
CMAKE_ASM_COMPILE_OPTIONS_EXTERNAL_TOOLCHAIN to "-gcc-toolchain ",
which newer versions of Clang reject in favor of "--gcc-toolchain=".

[1] Iaf4f20e97b173d463c7b4367903466fe8dc5cf64

Bug: https://github.com/android/ndk/issues/1623
Test: checkbuild.py
Change-Id: I957f47d570b566f20003eb11995763daa88ab94e