Fix bug in libshaderc build.
The missing comma meant that these strings were being concatenated.
Amazingly this was actually caught by a combination of reformatting
(which reflowed these onto the same line) and pylint (which has a
warning for useless concatenations on the same line).

This module is intentionally permissive with missing files according
to a comment below. I haven't changed that, but maybe it should be
changed.

Test: treehugger
Bug: None
Change-Id: Ia44bb8ce004a97e8e8300557f139e5eaa27e608c