Simplify module registration.
I always forget to update ALL_MODULES when I add a new module. A
decorator that is used at the class definition is hopefully harder to
forget. If that still doesn't help we can make this automatic with a
metaclass but that's almost always overkill (and parent classes will
need some special behavior to avoid being registered).

Test: treehugger
Bug: None
Change-Id: Ifd2c89efbc01339e886617803fe9c168f9206ee5