Add regression test for CMAKE_ANDROID_ARM_MODE.
We should not be setting this for anything but armeabi-v7a. We don't
by default, but if the user passes `-DANDROID_ARM_MODE` when building
for another ABI we'll still set this, and CMake 3.22 will emit an
error in that case. The legacy toolchain would ignore
`ANDROID_ARM_MODE` for non-armeabi-v7a ABIs.

Test: This is the test
Bug: https://gitlab.kitware.com/cmake/cmake/-/issues/23108
Change-Id: Ic0395fa7b632cf46980e14ea80221374799222b1