Batch test build results.
I thought I'd done this ages ago but apparently I only did it for test
**run** results. This generalizes the workqueue API a bit and makes
use of it during test builds. It's possible there's more to win back
here with things like push, but there are a much smaller number of
tasks there in the first place so it may not be significant.

This comes out to a ~33% reduction in test build time on my
workstation:

Before: 0:07:46
After: 0:05:15

In passing I also added a TypeVar to replace the Any result of the
workqueues, which caught some dead code.

Test: ./checkbuild.py pythonlint
Test: ./run_tests.py --rebuild
Bug: None
Change-Id: I9603e1b1519e9bdf9c11116cf9615a07dc70a22f