Check for distributed static libraries earlier.
Test enumeration checks for this before the tests get run, but we
should also check when we build the tests so we get an earlier
warning.

This also reworks some of the guts of the cmake/ndk-build test
builders so that the failed command is included in the output for
easier debugging.

Test: ran tsan_smoke test on windows
Bug: None
Change-Id: I580c4374ae694ba8bea6ed2b511831f91d673c24