Rework Python development environment checks.
We don't need the bootstrapping step any more because the build
servers already use our prebuilt Python, and local builds use Poetry
to accomplish the same. I've removed the bootstrapping step and
replaced it with verification of the Python environment. This checks
that the version of Python being used matches our prebuilts, that a
Poetry environment is being used if Poetry is installed, and that the
user site packages directory is not in sys.path.

Test: ./checkbuild.py and ./run_tests.py with and without poetry env
Bug: None
Change-Id: I5638d1aa5c6fb15c2d7591b7712ad6fe60139c08